fix: various bug fixes to client-side deserialization lesson
* awk

* awk

remove "in"

* awk

ambiguous "it"

* awk

* id --> ID

* sha --> SHA

* awk

run on

* singular "derivation"

it's a little awk, but is is correct.  Maybe an alt would be to spell 
out the acronym? E.g., "the derivation of Program Derived Addresses 
is..." 
or
"Deriving PDAs is important because..."
...tl;dr it is the derivation (of x) which is (not are) important.

* run on

* awk

* awk

* awk

* to e is to error

consistent with other lesson code examples - we've been spelling it out 
in full